backend/requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.12,<4.0
django-cors-headers>=3.7,<4.0
pandas>=1.2,<2.0
numpy>=1.20,<2.0
Pillow>=8.0,<9.0
opencv-python>=4.5,<5.0
scikit-image>=0.18,<0.19
transformers>=4.0,<5.0
torch>=1.8,<2.0
scipy>=1.6,<2.0
matplotlib>=3.3,<4.0
seaborn>=0.11,<0.12